User Crud = Bcrypt for password

+++ DSCommerce/DSCommerce/Repositories/OrderRepository.cs
@@ -5,9 +5,9 @@ namespace DSCommerce.Repositories
     public interface OrderRepository
     {
         Task<List<OrderDTO>> FindAll();
-        Task<OrderDTO> FindById(int id);
+        Task<OrderDTO> FindById(long id);
         Task<OrderDTO> Insert(OrderDTO dto);
-        Task<OrderDTO> Update(OrderDTO dto, int id);
-        Task<bool> DeleteById(int id);
+        Task<OrderDTO> Update(OrderDTO dto, long id);
+        Task<bool> DeleteById(long id);
     }
 }
